Create LazyDog
Created a program in python that understands different commands and responds and/or performs an action. The robot can tell the weather, say different phrases based on mood, give the time and date, and send an email including various resources that are tailored towards engineers. I utilized the speech_recognition and Pyttsx3 packages to create the voice recognition and speech of the robot. I also used an API to get a real- time weather forecast of the Knoxville area.